Ability to add items to course: primitive version
This works, but needs plenty of work yet, mostly on eliminating the
lazy use of the JavaScript console for all notifications about the
various things that can go wrong (or indeed right). That stuff should
be done with `<Callout>`.

Also still to be done: a ton of internationalization.

Still: this fixes the core part of UICR-7.