Get reserve.itemId from resources, not props
I don't understand why, but in the `item` manifest entry's `path`
callback function, `props.resources` is undefined when the PUT mutator
is called (even though it works fine for GET). But the separate
`resources` parameter remains defined in both contexts, so I get get
the PUT to happen by using that instead.