Pass item though to reserve updater
It needs this so it can PUT the item to update its temporary location.

NOTE. This does not yet work reliably. Sometimes when navigating to
the edit-reserve page, the item is fetched (and so available for the
PUT), but other times it is not. I do not yet know what the conditions
are.