Item record is always loaded along with reserve
This seems to fix UICR-43

But I don't feel good about it because I don't understand _why_ is
fixes that problem.

All I have changed here is that the `path` function for the `item`
entry in the manifest now obtains the reserve record using
	get(props, 'resources.reserve.records.0');
instead of
	get(resources, 'reserve.records.0');
which you would think makes no difference.

Here is what I think _might_ be going on ...

By inspection, the path-function's callback parameters show the
reserve record showing up in `resources` one invocation before it does
in `props.resources`. When the path function is checking the former,
it returns a correct item-record URL which stripes-connect ignores for
some reason, and then when the props version becomes available and we
return the same URL, stripes-connect sees that the URL is unchanged
from the last one that was returned, and uses its empty cache. But
when the path function is checking props it delays the return of the
correct path by one invocation, so that when it does so
stripes-connect seens the URL having changed and makes the request to
fetch the record.

NOTE that this is all supposition, and even if it's correct it tells
us nothing about why stripes-connect doesn't fetch the item record
when the `resources` parameter says the reserve is ready. So if my
guess is right, what we're seeing here is a bug in stripes-connect and
the present change is just a workaround. But I would not place any
confidence in that conclusion.