GET and PUT both work for item associated with reserve
This is deeply, deeply unsatisfying. To make this work I have had to
provide similar but different `path` functions for the GET and PUT
cases on the `item` manifest entry. The former accesses the reserve
object using the callback function's `resource` parameter, the latter
using its `props` parameter. I have no remotely sane rationale for why
this works, but it does -- and no other combination does.