List courses in reserve-editing course
A reserve belongs to a course-listing rather than to a single course,
so we can't just display The Course that the reserve belongs
to. Instead the form now includes a card which lists all the courses
that are cross-listed together in the course-listing that owns thw
reserve. Each course-name is a link to the course itself.

Fixes part of UICR-10.

Each course line