Correct (as far as I can tell) query-index names
I think this means that the search (as opposed to filter) part of
UICR-30 is done on the front-end -- but it's not actually _working_
yet, due to absence of back-end support for the queries.

(Or perhaps the back-end does support the necessary queries, but they
are specified in a different way from what I currently have. If so,
Kurt can put me right.)