Deleting an instructor from a course forces a reload/rerender
This was surprisingly awkward to do, and even now I wonder if there is
a better way that I somehow missed. I have added a local resource to
the course-display component, and included its value as an unused
parameter in the URL generated for fetching course data. I pass that
local resource's mutator down into the component that deletes the
instructor, and once that deletion is accomplished it mutates the
local resource to change its value (to the number of remaining
instructors). Stripes-connect spots that the URL of the course itself
has changed and reloads it, resulting in a new render.

It works, but it feels rather around-the-houses.