Completely new approach to fixing UICR-57.
Now, instead of petitioning stripes-connect to use the `itemId` in the
reserve that it fetches as the key to fetch an the item associated
with the reserve, we just pass the item-ID right in with the other
three IDs as part of the route. Admittedly this gives us ugly UI URLs
such as
http://localhost:3000/cr/reserves/736cb736-6ca6-4bc1-a046-3648314a5db7/5d849b08-4000-4aee-a80c-619a25521563/9aeee19a-5d9c-41e7-a021-
5f52994951a4/9a243c3a-3702-11ea-92ee-271862221368/edit
but they were already three quarters as ugly as that: they already had
course-listing ID, course ID and reserve ID, so we've just added item
ID.

I am much more confident about this approach to fixing the
looong-running complex of issues that first manifested as UICR-43, was
fixed in a bewildering way (different manifest-path functions for GET
and PUT) and has now popped up again as UICR-57. Everything is
simpler (apart form the URL itself). As a bonus, this version will be
slightly faster, as the reserve and item will load in parallel rather
than in serial.