Do not PUT to item when editing a reserve
Previously this was necessary to update the `temporaryLocationId`.
But as of mod-courses v1.0.1 (as part of the MODCR-31 work), it's
possible to set `copiedItem.temporaryLocationId` in the reserve, and
the new value will be copied into the item.

This change is not only desirable, but also necessary since the
additional PUT to the item was overwriting other changes that
mod-courses was making to the item when the reserve was PUT.