rename script entries that break platform-complete
When the `test` and `postinstall` hooks that are herein renamed with the
`-UNUSED` suffix are present, it causes the `platform-complete#snapshot`
build to fail in a very odd way. Specifically, it causes ui-courses to
not see any of the modules hoisted/provided by the platform. When it
comes to those modules ui-courses is supposed to receive via
peer-dependencies, this is problematic because they are not allowed to
be duplicated in the bundle.

I have no explanation for _why_ this is happening. I can only tell
you that when these two lines are present it _is_ happening and it
breaks the build. Running theory: `test` and `postinstall` are somehow
special to yarn, for definitions of "special" where it means "The battle
of stripes, 2020 ... your build turned upside down!"