Clarify the state of BigTest development
Based on the READMEs of the archived BigTest project repositories archives, the conclusion that BigTest is no longer maintained is understandable, but not true.

In fact, all of the original project repositories were combined into a single [monorepo][1] in order to make development more friction free. In order to clearly indicate this, the all of the archived projects have been updated to include links to their new home. 

[1]: http://github.com/thefrontside/bigtest