Add "kill" script to clean up lingering servers
For reasons I do not yet understand, the PID-saving and killing code
in the `yarn test` and `yarn regenerate` scripts does not obtain the
right PIDs: I suspect Node is fork-and-execing. As a result, the
processes do not get cleanly killed at the end of the run.

`yarn kill` kills any processes listening on ports 3001 and 3002.