UICR-116: Removed metadata when viewing course
Metadata is handled confusingly on our records. It's tracked
separately for course, courseListing, instructor, and reserve
records. This means that to get the last time the "Course" was
updated, we need to run through all of that metadata and see
which is latest. Sure, possible, but the final wrinkle is actually
when an instructor or reserve is removed from a courseListing.
In that case, the course or courseListing's metadata are not
updated, so the lastUpdated date can actually /go backwards
in time./ Woof. So, we're backing out this functionality until the
backend changes the way it handles metadata.