UICR-161 use local notes routes
Use local routes (i.e. `/cr/...`) for editing and viewing notes instead
of pointing to `/users` some of the time. This led to a very confusing
error message from react-intl about calling `<FormatMessage>` without
and `id` attribute. The SSC note-edit and note-view components rely on
`<ReferredRecord>` to display the note type but the courses and users
routes that render the notes only pass in keys corresponding to the
types they expect to handle. Thus, passing in a "courses" note with
`type: "course"` to a users route that only contains instructions for
how to format `type: "user"` causes an NPE.

Refs UICR-161