UICR-155 @babel/eslint-parser (#216)
* Replace `babel-eslint` with `@babel/eslint-parser`

Fixes UICR-155.

**NOTE** that I have TEMPORARILY disabled ESLint rejecting prop-type `array`

The upgrade from an older `eslint-config-stripes` under `babel-eslint` to a newer `eslint-config-stripes` under `@babel/eslint-parser` has introduced new ESLint errors, all of them rejection of the underspecified prop-type `array`.

The correct fix is to upgrade all the uses of `array` to `arrayOf`, and I will do that soon. But I can't do it right now because the only way to check I've got the member-types right is to actually run the Course Reserves app and check there are no prop-types warnings in the console. And I can't do that now because we don't have the MODCR-101 fix in play yet.