When deleting a course, also delete its courselisting  (#274)
This is intended to prevent orphan course-listings from lying around, preventing the deletion of terms that are still in use by them (see UICR-203). If the deletion of course-listing fails, we ignore that failure, as it's likely due to the existence of another course using it.

This works so far as it goes, but it turns out that course-listing deletion also fails if there are instructors assigned, and that these need to be deleted first. Since I'm not sure whether that ought to be the UI's responsibility, I'm ignoring it for now and committing the present version.

Fixes UICR-206.